In this stage, your program should satisfy the following conditions:
Every line of the input contains the person's name and the action this person performed (sent, joined, left, connected, and disconnected).
The person's name is a single word; there are no character restrictions for the person’s name.
The sent messages match the pattern "Name sent Message".
The messages in the output match the pattern "Name: Message".
All other types of messages are not shown in the output.
Your program will receive only correct raw messages.