The following code is supposed to find the general sum of two inclusive integer ranges in parallel. Unfortunately, this code doesn't work correctly. Try and fix it. Read the comments for a better understanding.
Additional explanations: the implemented class RangeSummator extends Thread and overrides the method run.
It sums the range passed to the constructor (left and right limits as integers). The class has a getter getResult()
that returns the calculated sum or 0 if the thread has not calculated the sum yet.

Note: you do not need to write the RangeSummator class. It will be added to your solution automatically.